requirements.build.txt
click==8.0.3
peewee==3.14.9
python-telegram-bot==13.11
